+++ src/rcpp_move_meta.h
@@ -1,7 +1,7 @@
 //#ifndef RCPP_MOVE_META
 //#define RCPP_MOVE_META
 
-Rcpp::List rcpp_move_meta(Rcpp::List fishpop, Rcpp::NumericVector stationary_values,
+Rcpp::List rcpp_move_meta(Rcpp::List fishpop, Rcpp::NumericVector residence_values,
                           int pop_n_sum, Rcpp::IntegerVector id_attr, Rcpp::IntegerVector id_meta,
                           Rcpp::NumericVector extent);