+++ tests/testthat/test-create_attributes.R
@@ -36,7 +36,7 @@ test_that("create_attributes has correct dimeonsions", {
   expect_equal(object = ncol(attributes_matrix), expected = 3)
 
   expect_equal(object = colnames(attributes_matrix),
-               expected = c("id", "stationary", "reserves_thres"))
+               expected = c("id", "residence", "reserves_thres"))
 
 })