+++ NEWS.md
@@ -3,6 +3,7 @@
 * Renamed `simulate_meta` to `run_meta`
 * Adding plotting method for `summarize = FALSE`
 * Adding `get_global_range`
+* Renamed `stationary` to `residence`
 
 # meta.arrR 0.2.0
 * First running framework